refactor(config): split config modules
