--- src/config.rs
@@ -1,1299 +1,0 @@
-use std::fs;
-use std::path::{Path, PathBuf};
-use std::time::Duration;
-
-use serde::Deserialize;
-
-use crate::app::scale::{ZOOM_MAX, ZOOM_MIN};
-use crate::app::{PageLayoutMode, SpreadCoverPolicy, SpreadDirection};
-use crate::error::{AppError, AppResult};
-use crate::input::keymap::build_builtin_sequence_registry;
-use crate::input::sequence::{DEFAULT_SEQUENCE_TIMEOUT, SequenceRegistry};
-
-#[derive(Debug, Clone, PartialEq, Default)]
-pub struct Config {
-    pub render: RenderConfig,
-    pub cache: CacheConfig,
-    pub view: ViewConfig,
-    pub input: InputConfig,
-    pub watch: WatchConfig,
-}
-
-#[derive(Debug, Clone, Deserialize, PartialEq)]
-#[serde(default)]
-pub struct RenderConfig {
-    pub worker_threads: usize,
-    pub input_poll_timeout_idle_ms: u64,
-    pub input_poll_timeout_busy_ms: u64,
-    pub prefetch_pause_ms: u64,
-    pub prefetch_tick_ms: u64,
-    pub pending_redraw_interval_ms: u64,
-    pub prefetch_dispatch_budget_per_tick: usize,
-    pub max_render_scale: f32,
-}
-
-impl Default for RenderConfig {
-    fn default() -> Self {
-        Self {
-            worker_threads: 3,
-            input_poll_timeout_idle_ms: 16,
-            input_poll_timeout_busy_ms: 8,
-            prefetch_pause_ms: 120,
-            prefetch_tick_ms: 8,
-            pending_redraw_interval_ms: 33,
-            prefetch_dispatch_budget_per_tick: 6,
-            max_render_scale: 2.5,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Deserialize, PartialEq, Eq)]
-#[serde(default)]
-pub struct CacheConfig {
-    pub l1_memory_budget_mb: usize,
-    pub l2_memory_budget_mb: usize,
-    pub l1_max_entries: usize,
-    pub l2_max_entries: usize,
-}
-
-impl Default for CacheConfig {
-    fn default() -> Self {
-        Self {
-            l1_memory_budget_mb: 512,
-            l2_memory_budget_mb: 64,
-            l1_max_entries: 128,
-            l2_max_entries: 96,
-        }
-    }
-}
-
-impl CacheConfig {
-    const MEBIBYTE: usize = 1024 * 1024;
-
-    pub fn l1_memory_budget_bytes(&self) -> usize {
-        self.l1_memory_budget_mb
-            .saturating_mul(Self::MEBIBYTE)
-            .max(1)
-    }
-
-    pub fn l2_memory_budget_bytes(&self) -> usize {
-        self.l2_memory_budget_mb
-            .saturating_mul(Self::MEBIBYTE)
-            .max(1)
-    }
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct ViewConfig {
-    pub initial_page: usize,
-    pub initial_zoom: f32,
-    pub initial_layout: PageLayoutMode,
-    pub spread_direction: SpreadDirection,
-    pub spread_cover: SpreadCoverPolicy,
-}
-
-impl Default for ViewConfig {
-    fn default() -> Self {
-        Self {
-            initial_page: 1,
-            initial_zoom: 1.0,
-            initial_layout: PageLayoutMode::Single,
-            spread_direction: SpreadDirection::Ltr,
-            spread_cover: SpreadCoverPolicy::Paired,
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct InputConfig {
-    pub sequence_timeout_ms: u64,
-}
-
-impl Default for InputConfig {
-    fn default() -> Self {
-        Self {
-            sequence_timeout_ms: DEFAULT_SEQUENCE_TIMEOUT.as_millis() as u64,
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct WatchConfig {
-    pub enabled: bool,
-    pub poll_interval_ms: u64,
-    pub settle_delay_ms: u64,
-}
-
-impl Default for WatchConfig {
-    fn default() -> Self {
-        Self {
-            enabled: false,
-            poll_interval_ms: 250,
-            settle_delay_ms: 500,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Deserialize, Default, PartialEq)]
-#[serde(default)]
-struct RawConfig {
-    render: Option<RawRenderConfig>,
-    cache: Option<RawCacheConfig>,
-    view: Option<RawViewConfig>,
-    input: Option<RawInputConfig>,
-    watch: Option<RawWatchConfig>,
-}
-
-#[derive(Debug, Clone, Deserialize, Default, PartialEq)]
-#[serde(default)]
-struct RawRenderConfig {
-    worker_threads: Option<usize>,
-    input_poll_timeout_idle_ms: Option<u64>,
-    input_poll_timeout_busy_ms: Option<u64>,
-    prefetch_pause_ms: Option<u64>,
-    prefetch_tick_ms: Option<u64>,
-    pending_redraw_interval_ms: Option<u64>,
-    prefetch_dispatch_budget_per_tick: Option<usize>,
-    max_render_scale: Option<f32>,
-}
-
-#[derive(Debug, Clone, Deserialize, Default, PartialEq, Eq)]
-#[serde(default)]
-struct RawCacheConfig {
-    l1_memory_budget_mb: Option<usize>,
-    l2_memory_budget_mb: Option<usize>,
-    l1_max_entries: Option<usize>,
-    l2_max_entries: Option<usize>,
-}
-
-#[derive(Debug, Clone, Deserialize, Default, PartialEq)]
-#[serde(default)]
-struct RawViewConfig {
-    initial_page: Option<usize>,
-    initial_zoom: Option<f32>,
-    initial_layout: Option<String>,
-    spread_direction: Option<String>,
-    spread_cover: Option<String>,
-}
-
-#[derive(Debug, Clone, Deserialize, Default, PartialEq, Eq)]
-#[serde(default)]
-struct RawInputConfig {
-    sequence_timeout_ms: Option<u64>,
-}
-
-#[derive(Debug, Clone, Deserialize, Default, PartialEq, Eq)]
-#[serde(default)]
-struct RawWatchConfig {
-    enabled: Option<bool>,
-    poll_interval_ms: Option<u64>,
-    settle_delay_ms: Option<u64>,
-}
-
-#[derive(Debug, Clone, Default, PartialEq)]
-pub struct AppOptions {
-    pub render: RenderOptions,
-    pub cache: CacheOptions,
-    pub view: ViewOptions,
-    pub input: InputOptions,
-    pub watch: WatchOptions,
-}
-
-impl AppOptions {
-    pub fn merge(mut self, next: Self) -> Self {
-        self.render = self.render.merge(next.render);
-        self.cache = self.cache.merge(next.cache);
-        self.view = self.view.merge(next.view);
-        self.input = self.input.merge(next.input);
-        self.watch = self.watch.merge(next.watch);
-        self
-    }
-}
-
-impl From<Config> for AppOptions {
-    fn from(config: Config) -> Self {
-        Self {
-            render: RenderOptions {
-                worker_threads: Some(config.render.worker_threads),
-                input_poll_timeout_idle_ms: Some(config.render.input_poll_timeout_idle_ms),
-                input_poll_timeout_busy_ms: Some(config.render.input_poll_timeout_busy_ms),
-                prefetch_pause_ms: Some(config.render.prefetch_pause_ms),
-                prefetch_tick_ms: Some(config.render.prefetch_tick_ms),
-                pending_redraw_interval_ms: Some(config.render.pending_redraw_interval_ms),
-                prefetch_dispatch_budget_per_tick: Some(
-                    config.render.prefetch_dispatch_budget_per_tick,
-                ),
-                max_render_scale: Some(config.render.max_render_scale),
-            },
-            cache: CacheOptions {
-                l1_memory_budget_mb: Some(config.cache.l1_memory_budget_mb),
-                l2_memory_budget_mb: Some(config.cache.l2_memory_budget_mb),
-                l1_max_entries: Some(config.cache.l1_max_entries),
-                l2_max_entries: Some(config.cache.l2_max_entries),
-            },
-            view: ViewOptions {
-                initial_page: Some(config.view.initial_page),
-                initial_zoom: Some(config.view.initial_zoom),
-                initial_layout: Some(config.view.initial_layout),
-                spread_direction: Some(config.view.spread_direction),
-                spread_cover: Some(config.view.spread_cover),
-            },
-            input: InputOptions {
-                sequence_timeout_ms: Some(config.input.sequence_timeout_ms),
-            },
-            watch: WatchOptions {
-                enabled: Some(config.watch.enabled),
-                poll_interval_ms: Some(config.watch.poll_interval_ms),
-                settle_delay_ms: Some(config.watch.settle_delay_ms),
-            },
-        }
-    }
-}
-
-impl RawConfig {
-    fn into_options(self) -> AppResult<AppOptions> {
-        Ok(AppOptions {
-            render: self.render.map(RenderOptions::from).unwrap_or_default(),
-            cache: self.cache.map(CacheOptions::from).unwrap_or_default(),
-            view: self
-                .view
-                .map(ViewOptions::try_from)
-                .transpose()?
-                .unwrap_or_default(),
-            input: self.input.map(InputOptions::from).unwrap_or_default(),
-            watch: self.watch.map(WatchOptions::from).unwrap_or_default(),
-        })
-    }
-}
-
-fn parse_page_layout_mode(value: &str) -> AppResult<PageLayoutMode> {
-    match value {
-        "single" => Ok(PageLayoutMode::Single),
-        "spread" => Ok(PageLayoutMode::Spread),
-        _ => Err(AppError::invalid_argument(format!(
-            "unknown view.initial_layout: {value}"
-        ))),
-    }
-}
-
-fn parse_spread_direction(value: &str) -> AppResult<SpreadDirection> {
-    match value {
-        "ltr" => Ok(SpreadDirection::Ltr),
-        "rtl" => Ok(SpreadDirection::Rtl),
-        _ => Err(AppError::invalid_argument(format!(
-            "unknown view.spread_direction: {value}"
-        ))),
-    }
-}
-
-fn parse_spread_cover(value: &str) -> AppResult<SpreadCoverPolicy> {
-    match value {
-        "paired" => Ok(SpreadCoverPolicy::Paired),
-        "cover" => Ok(SpreadCoverPolicy::Cover),
-        _ => Err(AppError::invalid_argument(format!(
-            "unknown view.spread_cover: {value}"
-        ))),
-    }
-}
-
-#[derive(Debug, Clone, Default, PartialEq)]
-pub struct RenderOptions {
-    pub worker_threads: Option<usize>,
-    pub input_poll_timeout_idle_ms: Option<u64>,
-    pub input_poll_timeout_busy_ms: Option<u64>,
-    pub prefetch_pause_ms: Option<u64>,
-    pub prefetch_tick_ms: Option<u64>,
-    pub pending_redraw_interval_ms: Option<u64>,
-    pub prefetch_dispatch_budget_per_tick: Option<usize>,
-    pub max_render_scale: Option<f32>,
-}
-
-impl RenderOptions {
-    fn merge(self, next: Self) -> Self {
-        Self {
-            worker_threads: next.worker_threads.or(self.worker_threads),
-            input_poll_timeout_idle_ms: next
-                .input_poll_timeout_idle_ms
-                .or(self.input_poll_timeout_idle_ms),
-            input_poll_timeout_busy_ms: next
-                .input_poll_timeout_busy_ms
-                .or(self.input_poll_timeout_busy_ms),
-            prefetch_pause_ms: next.prefetch_pause_ms.or(self.prefetch_pause_ms),
-            prefetch_tick_ms: next.prefetch_tick_ms.or(self.prefetch_tick_ms),
-            pending_redraw_interval_ms: next
-                .pending_redraw_interval_ms
-                .or(self.pending_redraw_interval_ms),
-            prefetch_dispatch_budget_per_tick: next
-                .prefetch_dispatch_budget_per_tick
-                .or(self.prefetch_dispatch_budget_per_tick),
-            max_render_scale: next.max_render_scale.or(self.max_render_scale),
-        }
-    }
-}
-
-impl From<RawRenderConfig> for RenderOptions {
-    fn from(raw: RawRenderConfig) -> Self {
-        Self {
-            worker_threads: raw.worker_threads,
-            input_poll_timeout_idle_ms: raw.input_poll_timeout_idle_ms,
-            input_poll_timeout_busy_ms: raw.input_poll_timeout_busy_ms,
-            prefetch_pause_ms: raw.prefetch_pause_ms,
-            prefetch_tick_ms: raw.prefetch_tick_ms,
-            pending_redraw_interval_ms: raw.pending_redraw_interval_ms,
-            prefetch_dispatch_budget_per_tick: raw.prefetch_dispatch_budget_per_tick,
-            max_render_scale: raw.max_render_scale,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Default, PartialEq, Eq)]
-pub struct CacheOptions {
-    pub l1_memory_budget_mb: Option<usize>,
-    pub l2_memory_budget_mb: Option<usize>,
-    pub l1_max_entries: Option<usize>,
-    pub l2_max_entries: Option<usize>,
-}
-
-impl CacheOptions {
-    fn merge(self, next: Self) -> Self {
-        Self {
-            l1_memory_budget_mb: next.l1_memory_budget_mb.or(self.l1_memory_budget_mb),
-            l2_memory_budget_mb: next.l2_memory_budget_mb.or(self.l2_memory_budget_mb),
-            l1_max_entries: next.l1_max_entries.or(self.l1_max_entries),
-            l2_max_entries: next.l2_max_entries.or(self.l2_max_entries),
-        }
-    }
-}
-
-impl From<RawCacheConfig> for CacheOptions {
-    fn from(raw: RawCacheConfig) -> Self {
-        Self {
-            l1_memory_budget_mb: raw.l1_memory_budget_mb,
-            l2_memory_budget_mb: raw.l2_memory_budget_mb,
-            l1_max_entries: raw.l1_max_entries,
-            l2_max_entries: raw.l2_max_entries,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Default, PartialEq)]
-pub struct ViewOptions {
-    pub initial_page: Option<usize>,
-    pub initial_zoom: Option<f32>,
-    pub initial_layout: Option<PageLayoutMode>,
-    pub spread_direction: Option<SpreadDirection>,
-    pub spread_cover: Option<SpreadCoverPolicy>,
-}
-
-impl ViewOptions {
-    fn merge(self, next: Self) -> Self {
-        Self {
-            initial_page: next.initial_page.or(self.initial_page),
-            initial_zoom: next.initial_zoom.or(self.initial_zoom),
-            initial_layout: next.initial_layout.or(self.initial_layout),
-            spread_direction: next.spread_direction.or(self.spread_direction),
-            spread_cover: next.spread_cover.or(self.spread_cover),
-        }
-    }
-}
-
-impl TryFrom<RawViewConfig> for ViewOptions {
-    type Error = AppError;
-
-    fn try_from(raw: RawViewConfig) -> Result<Self, Self::Error> {
-        Ok(Self {
-            initial_page: raw.initial_page,
-            initial_zoom: raw.initial_zoom,
-            initial_layout: raw
-                .initial_layout
-                .as_deref()
-                .map(parse_page_layout_mode)
-                .transpose()?,
-            spread_direction: raw
-                .spread_direction
-                .as_deref()
-                .map(parse_spread_direction)
-                .transpose()?,
-            spread_cover: raw
-                .spread_cover
-                .as_deref()
-                .map(parse_spread_cover)
-                .transpose()?,
-        })
-    }
-}
-
-#[derive(Debug, Clone, Default, PartialEq, Eq)]
-pub struct InputOptions {
-    pub sequence_timeout_ms: Option<u64>,
-}
-
-impl InputOptions {
-    fn merge(self, next: Self) -> Self {
-        Self {
-            sequence_timeout_ms: next.sequence_timeout_ms.or(self.sequence_timeout_ms),
-        }
-    }
-}
-
-impl From<RawInputConfig> for InputOptions {
-    fn from(raw: RawInputConfig) -> Self {
-        Self {
-            sequence_timeout_ms: raw.sequence_timeout_ms,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Default, PartialEq, Eq)]
-pub struct WatchOptions {
-    pub enabled: Option<bool>,
-    pub poll_interval_ms: Option<u64>,
-    pub settle_delay_ms: Option<u64>,
-}
-
-impl WatchOptions {
-    fn merge(self, next: Self) -> Self {
-        Self {
-            enabled: next.enabled.or(self.enabled),
-            poll_interval_ms: next.poll_interval_ms.or(self.poll_interval_ms),
-            settle_delay_ms: next.settle_delay_ms.or(self.settle_delay_ms),
-        }
-    }
-}
-
-impl From<RawWatchConfig> for WatchOptions {
-    fn from(raw: RawWatchConfig) -> Self {
-        Self {
-            enabled: raw.enabled,
-            poll_interval_ms: raw.poll_interval_ms,
-            settle_delay_ms: raw.settle_delay_ms,
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum ConfigFileSelection {
-    Default,
-    Path(PathBuf),
-    Disabled,
-}
-
-impl ConfigFileSelection {
-    pub fn load_options(&self) -> AppResult<AppOptions> {
-        match self {
-            Self::Default => load_default_app_options(),
-            Self::Path(path) => load_options_from_explicit_path(path),
-            Self::Disabled => Ok(AppOptions::default()),
-        }
-    }
-}
-
-pub fn load_default_app_options() -> AppResult<AppOptions> {
-    let Some(path) = default_config_path() else {
-        return Ok(AppOptions::default());
-    };
-    load_options_from_path_allow_missing(path)
-}
-
-pub fn load_options_from_explicit_path(path: impl AsRef<Path>) -> AppResult<AppOptions> {
-    read_options_from_path(path.as_ref(), MissingConfigPolicy::Error)
-}
-
-#[derive(Debug, Clone)]
-pub struct ResolvedAppOptions {
-    pub render: RenderPolicy,
-    pub view: ViewPolicy,
-    pub event_loop: EventLoopPolicy,
-    pub cache: CachePolicy,
-    pub input: InputPolicy,
-    pub watch: WatchPolicy,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub struct RenderPolicy {
-    pub worker_threads: usize,
-    pub max_render_scale: f32,
-}
-
-impl Default for RenderPolicy {
-    fn default() -> Self {
-        let render = RenderConfig::default();
-        Self {
-            worker_threads: render.worker_threads,
-            max_render_scale: render.max_render_scale,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct EventLoopPolicy {
-    pub input_poll_timeout_idle: Duration,
-    pub input_poll_timeout_busy: Duration,
-    pub prefetch_pause_after_input: Duration,
-    pub prefetch_tick_interval: Duration,
-    pub pending_redraw_interval: Duration,
-    pub prefetch_dispatch_budget_per_tick: usize,
-}
-
-impl Default for EventLoopPolicy {
-    fn default() -> Self {
-        let render = RenderConfig::default();
-        Self {
-            input_poll_timeout_idle: Duration::from_millis(render.input_poll_timeout_idle_ms),
-            input_poll_timeout_busy: Duration::from_millis(render.input_poll_timeout_busy_ms),
-            prefetch_pause_after_input: Duration::from_millis(render.prefetch_pause_ms),
-            prefetch_tick_interval: Duration::from_millis(render.prefetch_tick_ms),
-            pending_redraw_interval: Duration::from_millis(render.pending_redraw_interval_ms),
-            prefetch_dispatch_budget_per_tick: render.prefetch_dispatch_budget_per_tick,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct CachePolicy {
-    pub l1_memory_budget_mb: usize,
-    pub l2_memory_budget_mb: usize,
-    pub l1_max_entries: usize,
-    pub l2_max_entries: usize,
-}
-
-impl Default for CachePolicy {
-    fn default() -> Self {
-        let cache = CacheConfig::default();
-        Self {
-            l1_memory_budget_mb: cache.l1_memory_budget_mb,
-            l2_memory_budget_mb: cache.l2_memory_budget_mb,
-            l1_max_entries: cache.l1_max_entries,
-            l2_max_entries: cache.l2_max_entries,
-        }
-    }
-}
-
-impl CachePolicy {
-    const MEBIBYTE: usize = 1024 * 1024;
-
-    pub fn l1_memory_budget_bytes(&self) -> usize {
-        self.l1_memory_budget_mb
-            .saturating_mul(Self::MEBIBYTE)
-            .max(1)
-    }
-
-    pub fn l2_memory_budget_bytes(&self) -> usize {
-        self.l2_memory_budget_mb
-            .saturating_mul(Self::MEBIBYTE)
-            .max(1)
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub struct ViewPolicy {
-    pub initial_page_index: usize,
-    pub initial_zoom: f32,
-    pub initial_layout: PageLayoutMode,
-    pub spread_direction: SpreadDirection,
-    pub spread_cover: SpreadCoverPolicy,
-}
-
-impl Default for ViewPolicy {
-    fn default() -> Self {
-        let view = ViewConfig::default();
-        Self {
-            initial_page_index: view.initial_page - 1,
-            initial_zoom: view.initial_zoom,
-            initial_layout: view.initial_layout,
-            spread_direction: view.spread_direction,
-            spread_cover: view.spread_cover,
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct InputPolicy {
-    pub sequence_timeout: Duration,
-    pub sequence_registry: SequenceRegistry,
-}
-
-impl Default for InputPolicy {
-    fn default() -> Self {
-        Self {
-            sequence_timeout: DEFAULT_SEQUENCE_TIMEOUT,
-            sequence_registry: build_builtin_sequence_registry(),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct WatchPolicy {
-    pub enabled: bool,
-    pub poll_interval: Duration,
-    pub settle_delay: Duration,
-}
-
-impl Default for WatchPolicy {
-    fn default() -> Self {
-        let watch = WatchConfig::default();
-        Self {
-            enabled: watch.enabled,
-            poll_interval: Duration::from_millis(watch.poll_interval_ms),
-            settle_delay: Duration::from_millis(watch.settle_delay_ms),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Default)]
-pub struct AppOptionsResolver {
-    options: AppOptions,
-}
-
-impl AppOptionsResolver {
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    pub fn apply_options(mut self, options: AppOptions) -> Self {
-        self.options = self.options.merge(options);
-        self
-    }
-
-    pub fn resolve(self) -> ResolvedAppOptions {
-        resolve_options(self.options)
-    }
-}
-
-impl Default for ResolvedAppOptions {
-    fn default() -> Self {
-        AppOptionsResolver::new().resolve()
-    }
-}
-
-impl From<ResolvedAppOptions> for Config {
-    fn from(options: ResolvedAppOptions) -> Self {
-        Self {
-            render: RenderConfig {
-                worker_threads: options.render.worker_threads,
-                input_poll_timeout_idle_ms: options.event_loop.input_poll_timeout_idle.as_millis()
-                    as u64,
-                input_poll_timeout_busy_ms: options.event_loop.input_poll_timeout_busy.as_millis()
-                    as u64,
-                prefetch_pause_ms: options.event_loop.prefetch_pause_after_input.as_millis() as u64,
-                prefetch_tick_ms: options.event_loop.prefetch_tick_interval.as_millis() as u64,
-                pending_redraw_interval_ms: options.event_loop.pending_redraw_interval.as_millis()
-                    as u64,
-                prefetch_dispatch_budget_per_tick: options
-                    .event_loop
-                    .prefetch_dispatch_budget_per_tick,
-                max_render_scale: options.render.max_render_scale,
-            },
-            cache: CacheConfig {
-                l1_memory_budget_mb: options.cache.l1_memory_budget_mb,
-                l2_memory_budget_mb: options.cache.l2_memory_budget_mb,
-                l1_max_entries: options.cache.l1_max_entries,
-                l2_max_entries: options.cache.l2_max_entries,
-            },
-            view: ViewConfig {
-                initial_page: options.view.initial_page_index + 1,
-                initial_zoom: options.view.initial_zoom,
-                initial_layout: options.view.initial_layout,
-                spread_direction: options.view.spread_direction,
-                spread_cover: options.view.spread_cover,
-            },
-            input: InputConfig {
-                sequence_timeout_ms: options.input.sequence_timeout.as_millis() as u64,
-            },
-            watch: WatchConfig {
-                enabled: options.watch.enabled,
-                poll_interval_ms: options.watch.poll_interval.as_millis() as u64,
-                settle_delay_ms: options.watch.settle_delay.as_millis() as u64,
-            },
-        }
-    }
-}
-
-fn resolve_options(options: AppOptions) -> ResolvedAppOptions {
-    let render_defaults = RenderConfig::default();
-    let cache_defaults = CacheConfig::default();
-    let view_defaults = ViewConfig::default();
-    let watch_defaults = WatchConfig::default();
-
-    let worker_threads = options
-        .render
-        .worker_threads
-        .unwrap_or(render_defaults.worker_threads)
-        .max(1);
-    let input_poll_timeout_idle_ms = options
-        .render
-        .input_poll_timeout_idle_ms
-        .unwrap_or(render_defaults.input_poll_timeout_idle_ms)
-        .max(1);
-    let input_poll_timeout_busy_ms = options
-        .render
-        .input_poll_timeout_busy_ms
-        .unwrap_or(render_defaults.input_poll_timeout_busy_ms)
-        .max(1);
-    let prefetch_pause_ms = options
-        .render
-        .prefetch_pause_ms
-        .unwrap_or(render_defaults.prefetch_pause_ms)
-        .max(1);
-    let prefetch_tick_ms = options
-        .render
-        .prefetch_tick_ms
-        .unwrap_or(render_defaults.prefetch_tick_ms)
-        .max(1);
-    let pending_redraw_interval_ms = options
-        .render
-        .pending_redraw_interval_ms
-        .unwrap_or(render_defaults.pending_redraw_interval_ms)
-        .max(1);
-    let prefetch_dispatch_budget_per_tick = options
-        .render
-        .prefetch_dispatch_budget_per_tick
-        .unwrap_or(render_defaults.prefetch_dispatch_budget_per_tick)
-        .max(1);
-    let mut max_render_scale = options
-        .render
-        .max_render_scale
-        .unwrap_or(render_defaults.max_render_scale);
-    if !max_render_scale.is_finite() || max_render_scale < 1.0 {
-        max_render_scale = render_defaults.max_render_scale;
-    }
-    let sequence_timeout_ms = options
-        .input
-        .sequence_timeout_ms
-        .unwrap_or(DEFAULT_SEQUENCE_TIMEOUT.as_millis() as u64)
-        .max(1);
-    let initial_page_index = options
-        .view
-        .initial_page
-        .unwrap_or(view_defaults.initial_page)
-        .max(1)
-        - 1;
-    let mut initial_zoom = options
-        .view
-        .initial_zoom
-        .unwrap_or(view_defaults.initial_zoom);
-    if !initial_zoom.is_finite() || initial_zoom <= 0.0 {
-        initial_zoom = view_defaults.initial_zoom;
-    }
-    initial_zoom = initial_zoom.clamp(ZOOM_MIN, ZOOM_MAX);
-    let watch_poll_interval_ms = options
-        .watch
-        .poll_interval_ms
-        .unwrap_or(watch_defaults.poll_interval_ms)
-        .max(1);
-    let watch_settle_delay_ms = options
-        .watch
-        .settle_delay_ms
-        .unwrap_or(watch_defaults.settle_delay_ms)
-        .max(1);
-
-    ResolvedAppOptions {
-        render: RenderPolicy {
-            worker_threads,
-            max_render_scale,
-        },
-        view: ViewPolicy {
-            initial_page_index,
-            initial_zoom,
-            initial_layout: options
-                .view
-                .initial_layout
-                .unwrap_or(view_defaults.initial_layout),
-            spread_direction: options
-                .view
-                .spread_direction
-                .unwrap_or(view_defaults.spread_direction),
-            spread_cover: options
-                .view
-                .spread_cover
-                .unwrap_or(view_defaults.spread_cover),
-        },
-        event_loop: EventLoopPolicy {
-            input_poll_timeout_idle: Duration::from_millis(input_poll_timeout_idle_ms),
-            input_poll_timeout_busy: Duration::from_millis(input_poll_timeout_busy_ms),
-            prefetch_pause_after_input: Duration::from_millis(prefetch_pause_ms),
-            prefetch_tick_interval: Duration::from_millis(prefetch_tick_ms),
-            pending_redraw_interval: Duration::from_millis(pending_redraw_interval_ms),
-            prefetch_dispatch_budget_per_tick,
-        },
-        cache: CachePolicy {
-            l1_memory_budget_mb: options
-                .cache
-                .l1_memory_budget_mb
-                .unwrap_or(cache_defaults.l1_memory_budget_mb),
-            l2_memory_budget_mb: options
-                .cache
-                .l2_memory_budget_mb
-                .unwrap_or(cache_defaults.l2_memory_budget_mb),
-            l1_max_entries: options
-                .cache
-                .l1_max_entries
-                .unwrap_or(cache_defaults.l1_max_entries),
-            l2_max_entries: options
-                .cache
-                .l2_max_entries
-                .unwrap_or(cache_defaults.l2_max_entries),
-        },
-        input: InputPolicy {
-            sequence_timeout: Duration::from_millis(sequence_timeout_ms),
-            sequence_registry: build_builtin_sequence_registry(),
-        },
-        watch: WatchPolicy {
-            enabled: options.watch.enabled.unwrap_or(watch_defaults.enabled),
-            poll_interval: Duration::from_millis(watch_poll_interval_ms),
-            settle_delay: Duration::from_millis(watch_settle_delay_ms),
-        },
-    }
-}
-
-impl Config {
-    /// Loads the default config shape for compatibility with older call sites.
-    ///
-    /// Missing files fall back to defaults here because this API models optional
-    /// config discovery rather than explicit user selection.
-    pub fn load() -> AppResult<Self> {
-        let Some(path) = default_config_path() else {
-            return Ok(Self::default());
-        };
-        Self::load_from_path(path)
-    }
-
-    /// Loads the legacy resolved config shape from an optional path.
-    ///
-    /// Use `load_options_from_explicit_path` when a caller supplied a path that
-    /// must exist, such as `--config`.
-    pub fn load_from_path(path: impl AsRef<Path>) -> AppResult<Self> {
-        let options = read_options_from_path(path.as_ref(), MissingConfigPolicy::Default)?;
-        Ok(AppOptionsResolver::new()
-            .apply_options(options)
-            .resolve()
-            .into())
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum MissingConfigPolicy {
-    Default,
-    Error,
-}
-
-fn load_options_from_path_allow_missing(path: impl AsRef<Path>) -> AppResult<AppOptions> {
-    read_options_from_path(path.as_ref(), MissingConfigPolicy::Default)
-}
-
-fn read_options_from_path(path: &Path, missing: MissingConfigPolicy) -> AppResult<AppOptions> {
-    if !path.exists() {
-        return match missing {
-            MissingConfigPolicy::Default => Ok(AppOptions::default()),
-            MissingConfigPolicy::Error => Err(AppError::invalid_argument(format!(
-                "config path does not exist: {}",
-                path.display()
-            ))),
-        };
-    }
-    if !path.is_file() {
-        return Err(AppError::invalid_argument(format!(
-            "config path is not a regular file: {}",
-            path.display()
-        )));
-    }
-
-    let raw = fs::read_to_string(path).map_err(|source| {
-        AppError::io_with_context(source, format!("failed to read config: {}", path.display()))
-    })?;
-    let parsed = toml::from_str::<RawConfig>(&raw).map_err(|source| {
-        AppError::invalid_argument(format!(
-            "failed to parse config {}: {source}",
-            path.display()
-        ))
-    })?;
-    parsed.into_options()
-}
-
-pub fn default_config_path() -> Option<PathBuf> {
-    default_config_path_from_env(|key| std::env::var_os(key), Path::is_file)
-}
-
-fn default_config_path_from_env(
-    mut env_var: impl FnMut(&str) -> Option<std::ffi::OsString>,
-    is_file: impl Fn(&Path) -> bool,
-) -> Option<PathBuf> {
-    if let Some(explicit) = env_var("PVF_CONFIG_PATH")
-        && !explicit.is_empty()
-    {
-        return Some(PathBuf::from(explicit));
-    }
-
-    if let Some(xdg) = env_var("XDG_CONFIG_HOME")
-        && !xdg.is_empty()
-    {
-        let path = PathBuf::from(xdg).join("pvf").join("config.toml");
-        if is_file(&path) {
-            return Some(path);
-        }
-    }
-    if let Some(home) = env_var("HOME")
-        && !home.is_empty()
-    {
-        let path = PathBuf::from(home)
-            .join(".config")
-            .join("pvf")
-            .join("config.toml");
-        if is_file(&path) {
-            return Some(path);
-        }
-    }
-    if let Some(appdata) = env_var("APPDATA")
-        && !appdata.is_empty()
-    {
-        let path = PathBuf::from(appdata).join("pvf").join("config.toml");
-        if is_file(&path) {
-            return Some(path);
-        }
-    }
-    None
-}
-
-#[cfg(test)]
-mod tests {
-    use std::ffi::OsString;
-    use std::fs;
-    use std::path::PathBuf;
-    use std::process;
-    use std::time::{SystemTime, UNIX_EPOCH};
-
-    use std::time::Duration;
-
-    use crate::app::{PageLayoutMode, SpreadCoverPolicy, SpreadDirection};
-
-    use super::{
-        AppOptions, AppOptionsResolver, Config, RenderOptions, ViewOptions, WatchOptions,
-        default_config_path_from_env, load_options_from_explicit_path,
-    };
-
-    fn unique_temp_path(suffix: &str) -> PathBuf {
-        let nanos = SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .expect("clock should be after unix epoch")
-            .as_nanos();
-        let mut path = std::env::temp_dir();
-        path.push(format!("pvf_config_{suffix}_{}_{}", process::id(), nanos));
-        path
-    }
-
-    #[test]
-    fn load_from_path_returns_defaults_for_missing_file() {
-        let missing = unique_temp_path("missing.toml");
-        let config = Config::load_from_path(&missing).expect("missing config should fallback");
-        assert_eq!(config, Config::default());
-    }
-
-    #[test]
-    fn load_from_path_applies_partial_overrides_and_sanitizes() {
-        let path = unique_temp_path("custom.toml");
-        fs::write(
-            &path,
-            r#"
-            [render]
-            worker_threads = 0
-            input_poll_timeout_idle_ms = 0
-            input_poll_timeout_busy_ms = 0
-            prefetch_pause_ms = 0
-            prefetch_tick_ms = 0
-            pending_redraw_interval_ms = 0
-            prefetch_dispatch_budget_per_tick = 0
-            max_render_scale = 0.5
-
-            [cache]
-            l1_memory_budget_mb = 256
-
-            [view]
-            initial_page = 4
-            initial_zoom = 1.25
-            initial_layout = "spread"
-            spread_direction = "rtl"
-            spread_cover = "cover"
-
-            [input]
-            sequence_timeout_ms = 333
-
-            [watch]
-            enabled = true
-            poll_interval_ms = 125
-            settle_delay_ms = 250
-            "#,
-        )
-        .expect("config file should be written");
-
-        let config = Config::load_from_path(&path).expect("config should parse");
-        assert_eq!(config.render.worker_threads, 1);
-        assert_eq!(config.render.input_poll_timeout_idle_ms, 1);
-        assert_eq!(config.render.input_poll_timeout_busy_ms, 1);
-        assert_eq!(config.render.prefetch_pause_ms, 1);
-        assert_eq!(config.render.prefetch_tick_ms, 1);
-        assert_eq!(config.render.pending_redraw_interval_ms, 1);
-        assert_eq!(config.render.prefetch_dispatch_budget_per_tick, 1);
-        assert_eq!(config.render.max_render_scale, 2.5);
-        assert_eq!(config.cache.l1_memory_budget_mb, 256);
-        assert_eq!(config.cache.l2_memory_budget_mb, 64);
-        assert_eq!(config.cache.l1_max_entries, 128);
-        assert_eq!(config.cache.l2_max_entries, 96);
-        assert_eq!(config.view.initial_page, 4);
-        assert_eq!(config.view.initial_zoom, 1.25);
-        assert_eq!(config.view.initial_layout, PageLayoutMode::Spread);
-        assert_eq!(config.view.spread_direction, SpreadDirection::Rtl);
-        assert_eq!(config.view.spread_cover, SpreadCoverPolicy::Cover);
-        assert_eq!(config.input.sequence_timeout_ms, 333);
-        assert!(config.watch.enabled);
-        assert_eq!(config.watch.poll_interval_ms, 125);
-        assert_eq!(config.watch.settle_delay_ms, 250);
-
-        fs::remove_file(&path).expect("config file should be removed");
-    }
-
-    #[test]
-    fn load_from_path_ignores_legacy_keymap_section() {
-        let path = unique_temp_path("legacy-keymap.toml");
-        fs::write(
-            &path,
-            r#"
-            [keymap]
-            preset = "emacs"
-
-            [cache]
-            l1_max_entries = 42
-            "#,
-        )
-        .expect("config file should be written");
-
-        let config = Config::load_from_path(&path).expect("config should parse");
-        assert_eq!(config.cache.l1_max_entries, 42);
-
-        fs::remove_file(&path).expect("config file should be removed");
-    }
-
-    #[test]
-    fn load_options_from_explicit_path_preserves_partial_source_absence() {
-        let path = unique_temp_path("partial-options.toml");
-        fs::write(
-            &path,
-            r#"
-            [cache]
-            l1_max_entries = 42
-            "#,
-        )
-        .expect("config file should be written");
-
-        let options = load_options_from_explicit_path(&path).expect("options should parse");
-        assert_eq!(options.cache.l1_max_entries, Some(42));
-        assert_eq!(options.cache.l1_memory_budget_mb, None);
-        assert_eq!(options.render.worker_threads, None);
-        assert_eq!(options.render.max_render_scale, None);
-        assert_eq!(options.view.initial_page, None);
-        assert_eq!(options.input.sequence_timeout_ms, None);
-        assert_eq!(options.watch.enabled, None);
-
-        fs::remove_file(&path).expect("config file should be removed");
-    }
-
-    #[test]
-    fn load_options_from_explicit_path_reads_view_input_and_watch_sections() {
-        let path = unique_temp_path("view-input-watch-options.toml");
-        fs::write(
-            &path,
-            r#"
-            [view]
-            initial_page = 8
-            initial_zoom = 1.5
-            initial_layout = "spread"
-            spread_direction = "rtl"
-            spread_cover = "cover"
-
-            [input]
-            sequence_timeout_ms = 750
-
-            [watch]
-            enabled = true
-            poll_interval_ms = 100
-            settle_delay_ms = 200
-            "#,
-        )
-        .expect("config file should be written");
-
-        let options = load_options_from_explicit_path(&path).expect("options should parse");
-        assert_eq!(options.view.initial_page, Some(8));
-        assert_eq!(options.view.initial_zoom, Some(1.5));
-        assert_eq!(options.view.initial_layout, Some(PageLayoutMode::Spread));
-        assert_eq!(options.view.spread_direction, Some(SpreadDirection::Rtl));
-        assert_eq!(options.view.spread_cover, Some(SpreadCoverPolicy::Cover));
-        assert_eq!(options.input.sequence_timeout_ms, Some(750));
-        assert_eq!(options.watch.enabled, Some(true));
-        assert_eq!(options.watch.poll_interval_ms, Some(100));
-        assert_eq!(options.watch.settle_delay_ms, Some(200));
-
-        fs::remove_file(&path).expect("config file should be removed");
-    }
-
-    #[test]
-    fn load_options_from_explicit_path_rejects_unknown_view_values() {
-        let path = unique_temp_path("bad-view-options.toml");
-        fs::write(
-            &path,
-            r#"
-            [view]
-            initial_layout = "grid"
-            "#,
-        )
-        .expect("config file should be written");
-
-        let err = load_options_from_explicit_path(&path).expect_err("config should be rejected");
-        assert!(
-            err.to_string().contains("unknown view.initial_layout"),
-            "unexpected error: {err}"
-        );
-
-        fs::remove_file(&path).expect("config file should be removed");
-    }
-
-    #[test]
-    fn load_options_from_explicit_path_rejects_missing_path() {
-        let missing = unique_temp_path("missing-explicit.toml");
-        let err = load_options_from_explicit_path(&missing)
-            .expect_err("explicit missing config should fail");
-        assert!(
-            err.to_string().contains("config path does not exist"),
-            "unexpected error: {err}"
-        );
-    }
-
-    #[test]
-    fn default_config_path_keeps_explicit_path_even_when_missing() {
-        let explicit = PathBuf::from("/tmp/pvf-explicit-config.toml");
-        let found = default_config_path_from_env(
-            |key| (key == "PVF_CONFIG_PATH").then(|| OsString::from(&explicit)),
-            |_| false,
-        );
-
-        assert_eq!(found, Some(explicit));
-    }
-
-    #[test]
-    fn default_config_path_falls_through_missing_implicit_locations() {
-        let xdg = PathBuf::from("/tmp/pvf-xdg-config");
-        let home = PathBuf::from("/tmp/pvf-home");
-        let appdata = PathBuf::from("/tmp/pvf-appdata");
-        let expected = home.join(".config").join("pvf").join("config.toml");
-
-        let found = default_config_path_from_env(
-            |key| match key {
-                "XDG_CONFIG_HOME" => Some(OsString::from(&xdg)),
-                "HOME" => Some(OsString::from(&home)),
-                "APPDATA" => Some(OsString::from(&appdata)),
-                _ => None,
-            },
-            |path| path == expected,
-        );
-
-        assert_eq!(found, Some(expected));
-    }
-
-    #[test]
-    fn resolver_applies_defaults_and_sanitizes_without_file_source() {
-        let options = AppOptions {
-            render: RenderOptions {
-                worker_threads: Some(0),
-                input_poll_timeout_idle_ms: Some(0),
-                input_poll_timeout_busy_ms: Some(0),
-                prefetch_pause_ms: Some(0),
-                prefetch_tick_ms: Some(0),
-                pending_redraw_interval_ms: Some(0),
-                prefetch_dispatch_budget_per_tick: Some(0),
-                max_render_scale: Some(0.5),
-            },
-            view: ViewOptions {
-                initial_page: Some(0),
-                initial_zoom: Some(10.0),
-                initial_layout: Some(PageLayoutMode::Spread),
-                spread_direction: Some(SpreadDirection::Rtl),
-                spread_cover: Some(SpreadCoverPolicy::Cover),
-            },
-            watch: WatchOptions {
-                enabled: Some(true),
-                poll_interval_ms: Some(0),
-                settle_delay_ms: Some(0),
-            },
-            ..AppOptions::default()
-        };
-
-        let resolved = AppOptionsResolver::new().apply_options(options).resolve();
-
-        assert_eq!(resolved.render.worker_threads, 1);
-        assert_eq!(
-            resolved.event_loop.input_poll_timeout_idle,
-            Duration::from_millis(1)
-        );
-        assert_eq!(
-            resolved.event_loop.input_poll_timeout_busy,
-            Duration::from_millis(1)
-        );
-        assert_eq!(
-            resolved.event_loop.prefetch_pause_after_input,
-            Duration::from_millis(1)
-        );
-        assert_eq!(
-            resolved.event_loop.prefetch_tick_interval,
-            Duration::from_millis(1)
-        );
-        assert_eq!(
-            resolved.event_loop.pending_redraw_interval,
-            Duration::from_millis(1)
-        );
-        assert_eq!(resolved.event_loop.prefetch_dispatch_budget_per_tick, 1);
-        assert_eq!(resolved.render.max_render_scale, 2.5);
-        assert_eq!(resolved.view.initial_page_index, 0);
-        assert_eq!(resolved.view.initial_zoom, 4.0);
-        assert_eq!(resolved.view.initial_layout, PageLayoutMode::Spread);
-        assert_eq!(resolved.view.spread_direction, SpreadDirection::Rtl);
-        assert_eq!(resolved.view.spread_cover, SpreadCoverPolicy::Cover);
-        assert!(resolved.watch.enabled);
-        assert_eq!(resolved.watch.poll_interval, Duration::from_millis(1));
-        assert_eq!(resolved.watch.settle_delay, Duration::from_millis(1));
-    }
-
-    #[test]
-    fn resolver_merges_later_options_over_earlier_options() {
-        let base = AppOptions {
-            render: RenderOptions {
-                worker_threads: Some(2),
-                ..RenderOptions::default()
-            },
-            ..AppOptions::default()
-        };
-        let override_options = AppOptions {
-            render: RenderOptions {
-                worker_threads: Some(4),
-                ..RenderOptions::default()
-            },
-            watch: WatchOptions {
-                enabled: Some(false),
-                ..WatchOptions::default()
-            },
-            ..AppOptions::default()
-        };
-        let base = base.merge(AppOptions {
-            watch: WatchOptions {
-                enabled: Some(true),
-                ..WatchOptions::default()
-            },
-            ..AppOptions::default()
-        });
-
-        let resolved = AppOptionsResolver::new()
-            .apply_options(base)
-            .apply_options(override_options)
-            .resolve();
-
-        assert_eq!(resolved.render.worker_threads, 4);
-        assert!(!resolved.watch.enabled);
-    }
-}
