+++ src/config/mod.rs
@@ -1,0 +1,17 @@
+mod file;
+mod options;
+mod policy;
+mod types;
+
+pub use file::{
+    ConfigFileSelection, default_config_path, load_default_app_options,
+    load_options_from_explicit_path,
+};
+pub use options::{
+    AppOptions, CacheOptions, InputOptions, RenderOptions, ViewOptions, WatchOptions,
+};
+pub use policy::{
+    AppOptionsResolver, CachePolicy, EventLoopPolicy, InputPolicy, RenderPolicy,
+    ResolvedAppOptions, ViewPolicy, WatchPolicy,
+};
+pub use types::{CacheConfig, Config, InputConfig, RenderConfig, ViewConfig, WatchConfig};
